Update Gnome configuration
Remove all application folders in the overview.

Remove GTK and icon themes as the vanilla theme (Adwaita) provides a
more consistent experience and is easier to maintain.

Add 'Quick Settings Audio Panel' extension which enables per-application
volume to be easily adjusted.

Replace 'Tray Icons Reloaded' with 'AppIndicator and KStatusNotifierItem
Support' as it appears to work better.